@@ -36,7 +36,7 @@
     "eslint-plugin-unused-imports": "^4.1.4",
     "lucide-react": "^0.396.0",
     "nanoid": "^5.0.7",
-    "next": "14.2.4",
+    "next": "^15.3.0",
     "next-themes": "^0.3.0",
     "pg": "^8.14.1",
     "react": "^18",
