@@ -1,4 +1,4 @@
-import { env } from "@/lib/env.js";
+import { env } from "@/lib/env";
 import { QdrantClient } from "@qdrant/js-client-rest";
 
 const QDRANT_URL = env.QDRANT_URL || "http://localhost:6333";
